Refactor inline wizard internals
