+++ src/renderer/components/InlineWizard/DocumentGenerationView/index.ts
@@ -1,0 +1,10 @@
+export { DocumentGenerationView } from './DocumentGenerationView';
+export type {
+	DocumentEditorProps,
+	DocumentGenerationViewProps,
+	DocumentSelectorProps,
+} from './types';
+export { DocumentEditor, DocumentSelector } from './components';
+export { AustinFactsDisplay } from '../AustinFactsDisplay';
+export { StreamingDocumentPreview } from '../StreamingDocumentPreview';
+export { GenerationCompleteOverlay } from '../GenerationCompleteOverlay';
